api/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
oauthlib
requests
python-dotenv
SQLAlchemy
